fix: operation with overwrite_existing_files set to false
The logger is an old component ... created at a time where the
idea of having an optional configuration step was not yet considered
a bad idea.

Hence, the component is created in STOPPED state. This causes dynamic
properties to update themselves right away. Therefore, setFile gets
*always* called and setting the file property creates a log file
even if the component is not yet running.

This was essentially making overwrite_existing_files = false unusable,
as the file would be created on property write, and then again in
the startHook. Moreover, it would create leftovers if some code would
set file multiple times while the component is not yet running.

This commit adds the relevant tests, and changes setFile to not do
anything if the component is not in RUNNING state. startHook is the
one that creates the file.

In addition, the tests now use proper temporary file/dir generation,
which avoids leftovers and crashes.